[#55] rename : component -> components
