+++ src/components/BottomSheetModal/ItemDatePickerModal.tsx
@@ -3,7 +3,7 @@ import BottomSheetModal from '.'
 import styled from '@emotion/styled'
 import CustomDatepicker from '../CustomDatePicker/CustomDatePicker'
 import { useRecoilState } from 'recoil'
-import { selectedDateState } from '../../pages/item/create/component/DatePlaceField'
+import { selectedDateState } from '../../pages/item/create/components/DatePlaceField'
 import ButtonLarge from '../ButtonLarge/ButtonLarge'
 import useModals from '../Modals/hooks/useModals'
 import { modals } from '../Modals'
