+++ src/components/BottomSheetModal/ItemNameInputModal.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from 'react'
 import styled from '@emotion/styled'
 import useModals from '../Modals/hooks/useModals'
 import { modals } from '../Modals'
-import { itemNameState } from '../../pages/item/create/component/BrandItemField/BrandItemField'
+import { itemNameState } from '../../pages/item/create/components/BrandItemField/BrandItemField'
 import { useRecoilState } from 'recoil'
 import BottomSheetModal from '.'
 import Header from '../Header/Header'
