+++ src/components/BottomSheetModal/ItemPlaceInputModal.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import BottomSheetModal from '.'
 import styled from '@emotion/styled'
 import { useRecoilState } from 'recoil'
-import { selectedPlaceState } from '../../pages/item/create/component/DatePlaceField'
+import { selectedPlaceState } from '../../pages/item/create/components/DatePlaceField'
 import ButtonLarge from '../ButtonLarge/ButtonLarge'
 import useModals from '../Modals/hooks/useModals'
 import { modals } from '../Modals'
