+++ src/pages/item/create/index.tsx
@@ -2,10 +2,10 @@ import React, { useState } from 'react'
 import BrandItemField, {
   itemNameState,
   selectedBrandState,
-} from './component/BrandItemField/BrandItemField'
+} from './components/BrandItemField/BrandItemField'
 import { useRecoilValue } from 'recoil'
-import DatePlaceField from './component/DatePlaceField/DatePlaceField'
-import PriceField, { itemPriceState } from './component/PriceField/PriceField'
+import DatePlaceField from './components/DatePlaceField/DatePlaceField'
+import PriceField, { itemPriceState } from './components/PriceField/PriceField'
 import { useNavigate } from 'react-router-dom'
 
 const ItemCreate = () => {
